+++ common/CompilerDetect.h
@@ -34,8 +34,8 @@
 #define MPT_CLANG_AT_LEAST(major,minor,patch)        (MPT_COMPILER_CLANG_VERSION >= MPT_COMPILER_MAKE_VERSION3((major),(minor),(patch)))
 #define MPT_CLANG_BEFORE(major,minor,patch)          (MPT_COMPILER_CLANG_VERSION <  MPT_COMPILER_MAKE_VERSION3((major),(minor),(patch)))
 
-#if MPT_CLANG_BEFORE(3,8,0)
-#error "clang version 3.8 required"
+#if MPT_CLANG_BEFORE(5,0,0)
+#error "clang version 5 required"
 #endif
 
 #if defined(__clang_analyzer__)
@@ -51,8 +51,8 @@
 #define MPT_GCC_AT_LEAST(major,minor,patch)          (MPT_COMPILER_GCC_VERSION >= MPT_COMPILER_MAKE_VERSION3((major),(minor),(patch)))
 #define MPT_GCC_BEFORE(major,minor,patch)            (MPT_COMPILER_GCC_VERSION <  MPT_COMPILER_MAKE_VERSION3((major),(minor),(patch)))
 
-#if MPT_GCC_BEFORE(5,1,0)
-#error "GCC version 5.1 required"
+#if MPT_GCC_BEFORE(7,1,0)
+#error "GCC version 7.1 required"
 #endif
 
 #elif defined(_MSC_VER)
@@ -139,25 +139,21 @@
 
 #if (__cplusplus >= 201703)
 #define MPT_CXX 17
-#elif (__cplusplus >= 201402)
-#define MPT_CXX 14
 #else
-#define MPT_CXX 14
+#define MPT_CXX 17
 #endif
 
 #elif MPT_COMPILER_MSVC
 
 #if (_MSVC_LANG >= 201703)
 #define MPT_CXX 17
-#elif (_MSVC_LANG >= 201402)
-#define MPT_CXX 14
 #else
-#define MPT_CXX 14
+#define MPT_CXX 17
 #endif
 
 #else
 
-#define MPT_CXX 14
+#define MPT_CXX 17
 
 #endif
 
@@ -305,11 +301,6 @@
 
 #if MPT_OS_DJGPP
 #define MPT_COMPILER_QUIRK_NO_WCHAR
-#endif
-
-
-#if MPT_GCC_BEFORE(6,1,0)
-#define MPT_COMPILER_QUIRK_NO_CONSTEXPR14_THROW
 #endif