+++ common/mptAlloc.cpp
@@ -39,28 +39,9 @@ namespace mpt
 
 
 
-#if MPT_CXX_AT_LEAST(17)
-#else
-void* align(std::size_t alignment, std::size_t size, void* &ptr, std::size_t &space) noexcept
-{
-	std::size_t offset = static_cast<std::size_t>(reinterpret_cast<std::uintptr_t>(ptr) & (alignment - 1));
-	if(offset != 0)
-	{
-		offset = alignment - offset;
-	}
-	if((space < offset) || ((space - offset) < size))
-	{
-		return nullptr;
-	}
-	ptr = static_cast<mpt::byte*>(ptr) + offset;
-	space -= offset;
-	return ptr;
-}
-#endif
-
 aligned_raw_memory aligned_alloc_impl(std::size_t size, std::size_t count, std::size_t alignment)
 {
-	#if MPT_CXX_AT_LEAST(17) && !defined(MPT_COMPILER_QUIRK_NO_ALIGNEDALLOC)
+	#if !defined(MPT_COMPILER_QUIRK_NO_ALIGNEDALLOC)
 		std::size_t space = count * size;
 		void* mem = std::aligned_alloc(alignment, space);
 		if(!mem)
@@ -86,7 +67,7 @@ aligned_raw_memory aligned_alloc_impl(std::size_t size, std::size_t count, std::
 				MPT_EXCEPTION_THROW_OUT_OF_MEMORY();
 			}
 			void* aligned_mem = mem;
-			void* aligned = mpt::align(alignment, size * count, aligned_mem, space);
+			void* aligned = std::align(alignment, size * count, aligned_mem, space);
 			if(!aligned)
 			{
 				MPT_EXCEPTION_THROW_OUT_OF_MEMORY();
@@ -107,7 +88,7 @@ aligned_raw_memory aligned_alloc_impl(std::size_t size, std::size_t count, std::
 
 void aligned_free(aligned_raw_memory raw)
 {
-	#if MPT_CXX_AT_LEAST(17) && !defined(MPT_COMPILER_QUIRK_NO_ALIGNEDALLOC)
+	#if !defined(MPT_COMPILER_QUIRK_NO_ALIGNEDALLOC)
 		std::free(raw.mem);
 	#elif MPT_COMPILER_MSVC || (defined(__clang__) && defined(_MSC_VER))
 		_aligned_free(raw.mem);