+++ mptrack/Mptrack.cpp
@@ -850,9 +850,6 @@ BOOL CTrackApp::InitInstanceEarly(CMPTCommandLineInfo &cmdInfo)
 		ExceptionHandler::useImplicitFallbackSEH = false;
 		ExceptionHandler::useExplicitSEH = true;
 		ExceptionHandler::handleStdTerminate = true;
-		#if MPT_CXX_BEFORE(17)
-			ExceptionHandler::handleStdUnexpected = true;
-		#endif
 		ExceptionHandler::handleMfcExceptions = true;
 		ExceptionHandler::debugExceptionHandler = true;
 	} else if(IsDebuggerPresent() || cmdInfo.m_noCrashHandler)
@@ -861,9 +858,6 @@ BOOL CTrackApp::InitInstanceEarly(CMPTCommandLineInfo &cmdInfo)
 		ExceptionHandler::useImplicitFallbackSEH = false;
 		ExceptionHandler::useExplicitSEH = false;
 		ExceptionHandler::handleStdTerminate = false;
-		#if MPT_CXX_BEFORE(17)
-			ExceptionHandler::handleStdUnexpected = false;
-		#endif
 		ExceptionHandler::handleMfcExceptions = false;
 		ExceptionHandler::debugExceptionHandler = false;
 	} else
@@ -872,9 +866,6 @@ BOOL CTrackApp::InitInstanceEarly(CMPTCommandLineInfo &cmdInfo)
 		ExceptionHandler::useImplicitFallbackSEH = true;
 		ExceptionHandler::useExplicitSEH = true;
 		ExceptionHandler::handleStdTerminate = true;
-		#if MPT_CXX_BEFORE(17)
-			ExceptionHandler::handleStdUnexpected = true;
-		#endif
 		ExceptionHandler::handleMfcExceptions = true;
 		ExceptionHandler::debugExceptionHandler = false;
 	}